src/feature_engineering/requirements.txt
gunicorn
flask
google-cloud-bigquery[pandas]
google-cloud-bigquery-storage
google-cloud-storage
numpy<2.0
pandas
pandas-ta
db-dtypes
python-dateutil
tenacity>=8.2.0
